removed sql_context as no longer needed (cleanup)
