+++ native/explorer/src/lib.rs
@@ -23,7 +23,6 @@ mod error;
 mod expressions;
 mod lazyframe;
 mod series;
-mod sql_context;
 
 pub use datatypes::{
     ExDataFrame, ExDataFrameRef, ExExpr, ExExprRef, ExLazyFrame, ExLazyFrameRef, ExSeries,
@@ -33,7 +32,6 @@ pub use datatypes::{
 pub use error::ExplorerError;
 use expressions::*;
 use series::*;
-pub use sql_context::*;
 
 mod atoms {
     rustler::atoms! {
